requirements.txt
beautifulsoup4
selenium 
pandas 
pyarrow 
Pillow 
requests